+++ CHANGES.md
@@ -1,7 +1,12 @@
 # Changes History
 
+1.1.1
+-----
+Add **__isset** method to DTO to have proper properties check on object
+Improve DTO class description
+
 1.1.0
-_____
+-----
 Add PartialUpdateDto
 
 1.0.9